feat(dashboard): implement Team vs Targets chart with gradient colors
## New Features
- Team vs Targets horizontal bar chart showing rep performance vs monthly targets
- Gradient color system based on achievement percentage:
  - Red: <50%
  - Red→Yellow: 51-70%
  - Yellow→Pale Green: 71-85%
  - Pale Green→Vivid Green: 86-100%+
- Filter out "Dismissed" and "Intercompany" reps from chart
- Dynamic chart height based on number of reps
- All sales rep initials visible on Y-axis

## Database Changes
- Fix sales_targets.rep_id column type from INTEGER to BIGINT
- Create team_vs_targets_month RPC function with BIGINT support
- Function returns: rep performance, targets, variance, achievement %, status

## Files Added
- supabase/migrations/20250114_team_vs_targets_rpc.sql
- supabase/migrations/20250114_fix_team_vs_targets_complete.sql
- scripts/fix-team-vs-targets.mjs

## Files Modified
- components/dashboard/dealer-sales-pulse.tsx
  - Added getAchievementColor() gradient function
  - Added interpolateColor() hex color blending
  - Team vs Targets chart with dynamic height
  - Filtered reps with trim() to handle trailing spaces
- supabase/migrations/20250113_create_sales_targets_table.sql
  - Changed rep_id from INTEGER to BIGINT

+++ supabase/migrations/20250113_create_sales_targets_table.sql
@@ -1,7 +1,7 @@
 -- Create sales_targets table to track monthly sales targets per rep
 CREATE TABLE IF NOT EXISTS sales_targets (
   id BIGSERIAL PRIMARY KEY,
-  rep_id INTEGER NOT NULL,
+  rep_id BIGINT NOT NULL,
   target_month TEXT NOT NULL, -- Format: YYYY-MM
   target_amount NUMERIC(12, 2) NOT NULL DEFAULT 200000.00,
   fiscal_year INTEGER NOT NULL,
